separate opengl from robotoverlord class
